+++ app/src/main/java/com/vssh/blesensordemo/ui/CooktopViewModel.java
@@ -17,22 +17,22 @@ public class CooktopViewModel extends AndroidViewModel{
     private GourmetSensor sensor;
     Food food;
     Cooktop cooktop;
-    MutableLiveData<Float> foodTemperature = new MutableLiveData<>();
+    MutableLiveData<Double> foodTemperature = new MutableLiveData<>();
 
     public CooktopViewModel(@NonNull Application application) {
         super(application);
 
         cooktop = new Cooktop();
         sensor = new GourmetSensor(this.getApplication());
-        foodTemperature.setValue(0f);
-        food = new Food(0.5f, 30, new Food.HeatTransferInterface() {
+        foodTemperature.setValue((double) 0);
+        food = new Food(0.5f, 30f, new Food.HeatTransferInterface() {
             @Override
             public int getInputHeat() {
                 return cooktop.getOutputAfterLosses();
             }
 
             @Override
-            public void setNewTemp(float temperature) {
+            public void setNewTemp(double temperature) {
                 foodTemperature.setValue(temperature);
             }
         }, sensor);
